我的求购 && 专利求购

+++ ruoyi-patent/src/main/java/com/ruoyi/patent/domain/GPatenLibraryLineUp.java
@@ -75,7 +75,7 @@ public class GPatenLibraryLineUp extends BaseEntity {
         this.lineUpNum = lineUpNum;
     }
 
-    public int getLineUpNum() {
+    public Long getLineUpNum() {
         return lineUpNum;
     }
